[#1865] Frontend: pin file title on scroll in the Code Panel (#1860)
Currently, users cannot hide/collapse a file after they finish reading
through it, without scrolling back to the top again.

Let us pin the title of the file (which includes the collapse button) so
that the user can collapse a file when/after reading through it instead
of only on the top. Let us also position the tooltips to show/hide a
file dynamically based on whether they are out of the viewport.