[#1887] Convert run-checks.sh into a Gradle task (#1907)
Developers are more likely to forget about the existence of 
environmental checks (even though they are important for passing CI) due
to the different abstraction levels compared to the other linters. They
are also unable to test these checks locally easily and have to rely on
GitHub Actions. 

Let's use Gradle as an abstraction for run-checks.sh. Bash scripts can 
be run for Linux and macOS, while batch scripts can be run for Windows.