Change serveTestReportInBackground workingDir
The working directory of serveTestReportInBackground is quite broad and
includes the entire build subdirectory. This results in tasks such as
checkstyle tasks being considered undeclared dependencies even when
they should not be.

Such unwanted dependencies will be flagged as errors in Gradle 8.

Let's move the working directory of serveTestReportInBackground into
a separate subdirectory and copy all the necessary files into this new
subdirectory.