[#1877] Show merge commits in the report (#1882)
Currently, merge commits are completely ignored during report generation
and analysis. This misrepresents the number of commits the user has made
and misses out on important information included in the merge commits,
such as tags.

Let's include merge commits in the report so that important information
in such commits is not missed out.