[#1915] Make ReportGenerator and its Reporters non-static (#1916)
Currently, ReportGenerator, as well as its corresponding Reporters 
(CommitsReporter and AuthorshipReporter) are static utility classes. 
From a simple lookthrough of git blame, I believe it is made static 
only for legacy reasons (when RepoSense was a smaller project).

These classes have to be made non-static if we want to parallelize/have
concurrent runs of RepoSense. For example, in ConfigSystemTest, each
RepoSense instance is run through the RepoSense.main method. In order 
to parallelize ConfigSystemTest, we would have to make these classes 
non-static so that there will not be any thread interference.

Let's make ReportGenerator and its corresponding Reporters non-static.