[#1934] Fix wrong author name for merged groups in Commits Panel (#1935)
Currently, the Commits Panel shows the wrong author name for merged
groups, as it just picks the first author in the group. This is
misrepresentative of the work of all authors who have contributed to
the commits shown.

This bug has occurred due to a typo in the code, where one of the
instances of 'zIsMerge' was not renamed to `zIsMerged` as required.

Let us fix this typo so that the dashboard works as expected.