[#1879] Remove unnecessary test mode logic in ConfigSystemTest (#1911)
Due to legacy reasons, the test mode flag was added in order to allow 
systemtest to pass. However, systemtest was previously failing because 
the json file were incorrect and should not have included the addition 
of authors not listed in author config files.

In reality, RepoSense's behavior is correct and thus do not require 
hasAuthorConfiguration to be set to false for all the systemtest.

Let's update the json files to more accurately match the actual 
behavior of RepoSense. Let's also remove the setting of 
hasAuthorConfiguration to false if test mode is enabled.