[#1936] Migrate store.js to TypeScript (#1937)
Currently, despite the addition of TypeScript support, the frontend is
still largely written in JavaScript. This results in a lack of type
safety and many complex objects being passed around as unknown types,
which may in turn lead to errors.

Let's migrate the store.js file to TypeScript, to get types for the all
the states that are stored within the Vuex store. This will provide a
good starting point to migrate the Vue files that access and mutate the
state of the store.