Remove Gradle wrapper version definition (#1951)
The gradleVersion property of the wrapper fixes the Gradle version used
by the wrapper. This is confusing and difficult to maintain as Gradle
gets upgraded.

Let's remove the definition of the gradleVersion property in the
wrapper. This results in the current version of Gradle itself being used
for the wrapper too.