[#1936] Migrate app.vue to TypeScript (#1938)
Currently, despite the addition of TypeScript support, the frontend is
still largely written in JavaScript. This results in a lack of type
safety and many complex objects being passed around as unknown types,
which may in turn lead to errors.

Let's migrate the root component, app.vue, to TypeScript. As this is the
entry point of the app, converting this file to TypeScript will enable
the conversion of other Vue components.