[#1537] Show authors for merged groups in Code Panel (#1939)
Currently, when the "merge group" option is used for repos with multiple
authors, the Code Panel does not show which line of code was written by
which author.

Let us use different colors for the code segment backgrounds to indicate
who the author of the segment is. Let us also add an author legend for
each file to easily see who edited that particular file, and identify
which lines of code were written by whom.