[#1909] Refactor CliArguments and subclasses for polymorphism (#1942)
Currently, ArgsParser::parse does the following (1) sanitizes and 
formats the inputs (2) decides whether the reposense run should be a 
simple view, a CLI run or a config file run. After returning an instance
of CliArguments, RepoSense main class decides based on the class type
how to construct the corresponding repoConfigurations and 
groupConfigurations. Once the repoConfigurations and groupConfigurations
are setup, it no longer matters whether it is a CLI run or config file
run. 

Refactoring the code will allow us to reduce the complexity of the 
ArgsParser::parse method and allow for a slightly more extensible 
approach to implementing possible new run configurations. For example, a
class can be added to decide the appropriate run configuration for the
CliArguments which then generates the appropriate Config files. 

Using the builder pattern for the construction of CliArguments will also
help to avoid a bloated constructor and optionally initialize fields. 

Let's encapsulate specific run configuration behavior (CLI or config) to
a specific class which makes it easier to maintain as well as adopt the
builder pattern for CliArguments.