[#1966] Add cypress test for merged groups contribution bars (#1970)
Currently, there are no tests to ensure that the length and the number 
of contribution bars for merged groups are correct. This resulted in 
issue #1959, where the contribution bar went missing, being undetected. 

Let's add a cypress test to check that the length and the number of 
contribution bars for merged groups are correct, even after refresh.