[#1971] Add Vue router (#1974)
Currently, the RepoSense frontend does not have a router as everything 
is in the root page. In the future, we might add other pages or other 
functionalities, such as widgets, which might require routing. Adding a 
router also allows for easier backward navigation, which can allow a 
user to click back to navigate to a previous view after changing 
something on the frontend. 

Let's add a Vue Router to enable routing and history management in the 
RepoSense frontend.