[#1936] Migrate c-summary-charts.vue to TypeScript (#1976)
Currently, despite the addition of TypeScript support, the frontend is
still largely written in JavaScript. This results in a lack of type
safety and many complex objects being passed around as unknown types,
which may in turn lead to errors.

Let's migrate one of the main components, c-summary-charts.vue, to
TypeScript. This will provide better type safety and reduce runtime
errors.