[#528] Allow multiple repos in one row for author-config.csv (#1961)
The same author may be working on several repositories to be analyzed.
In order to specify author information through author-config.csv, each
row has to be duplicated for each repository. This may be tedious for
the person generating the RepoSense report.

This change allows the repository location column in author-config.csv
to accept multiple repositories through the separator delimiter ";". It
also allows users to specify branches for each repository by using "~"
after the repository and separate multiple branches with "|", This
removes the need to duplicate the remaining columns' information in
multiple rows and reduce likelihood of errors in maintenance.

Let's allow each row in author_config.csv to contain multiple repos and
multiple branches for each repo to reduce row duplication.