[#1004] Standardize fonts for consistency across panels (#1979)
There are inconsistencies in the design between the different panels. 
One aspect that varies is the fonts used in terms of the font sizes and 
families. 

Let us standardize the font sizes across the Vue front-end using mixins.
Let us also set the font family for sans-serif to be Titillium Web and
monospace to be GitHub's default font.