[#1984] Frontend: fix CheckedFileTypes unable to be set as empty (#1985)
Currently, if "breakdown by file type" is selected and no file type is 
selected, if the page is refreshed/the URL is visited, all file types 
will be selected again. 

Let's fix it so the selected file types remain as configured by the user
in the URL, for both summary and code commit panels.