[#1908] Update how logger output is recorded (#1978)
Currently, setLogFolderLocation is called in ArgsParser which sets the
output folder location for all loggers. However, any logs before the
output folder is set do not get recorded in the final log file and only
show up in the standard output. This also potentially causes concurrency
issues, since the current structure causes ArgsParser::parse to have
side effects that affect LogsManager.

Let's output logging messages to a file reposense.log in the directory
where RepoSense is run and only move the reposense.log file to the
output directory at the end of the RepoSense run.