Add Graphviz installation step to CI
The Report Architecture image cannot be rendered automatically from
the puml due to Graphviz not being installed during the CI process.

Let's add the step to install Graphviz in the integration and GitHub
pages workflow.

By rendering the puml directly, this eliminates the need to manually
convert the puml to a png image.