[#1968] Incorrect dropdown box for config in UG: Customizing Report (#2009)
There is a bug in the UG where the assets syntax is incorrectly
displayed under the config section instead of the correct config
syntax. This could potentially confuse users.

Let's fix the bug by displaying the correct config syntax, and add a
link to the page that explains the format of config files, for user
convenience.