[#944] Fix wrong credit information inherited by annotated author (#2060)
AnnotatorAnalyzer only overwrites the author but not the credit
information, when an author tag is found. If the analyze authorship
flag is enabled, credit information based on the blame author will be
wrongly inherited by the annotated author.

Lets assign partial credit if the annotated author is not the same as
the blame author, and keep the analyzed credit information if the 2
are the same.