[#1986] Add cypress tests for renderFilterHash (#2017)
There were a few instances of URL parameters being incorrectly
loaded due to bugs in the code. 

These bugs were not able to be detected early due to a lack of 
tests for the renderFilterHash method.

Let's add tests to check that the URL parameters are loaded
correctly after the application reloads.