[#2001] Extract c-authorship-file component from views/c-authorship (#2096)
Many of the frontend files are difficult to navigate through due to the
large file sizes. As we add more features to the frontend, it is
getting harder to maintain. This is also very unfriendly towards new
contributors. 

Let us break down frontend files in a logical manner, starting with
extracting c-authorship-file from views/c-authorship.

This would provide a starting point towards further refactoring of the
frontend.