[#2109] Add search by tag functionality (#2167)
Add search by tag functionality

It can be useful to search author/repos by git tags.

Let's add this functionality to make it easier to filter by tags. This
commit also fixes a bug that existed in a previous version of the
feature which resulted in all users being shown to belong to same group.