[#2158] Add More Documentation for Title Component (#2159)
Add More Documentation for Title Component

The documentation for the title component is sparse. Adding more
details on how to use it will make it easier for users to add their own
titles to their reports.

Let's update the documentation to include a section for the title
component, and add the title component to the default view.