[#2148] Show tags on the ramp chart (#2163)
It can be useful to view all the tags of each repo that appears on the
ramp chart. Though this information is visible if the repo is merged
and the user click to see the breakdown of each commit, it can still be
useful to see this information upfront (for searching Ctrl+F perhaps).
This feature would be toggleable with a checkbox to reduce clutter.

Let's add this functionality to add convenience when searching through
RepoSense reports with tags.