[#944] Implement authorship analysis (#2140)
A line is credited to the author who last modified it.

Another author might have written the line initially and the current
author only modified it slightly. In such a case, the current author
gets credited for work that is not entirely done by him/her.

Let's analyze how similar a line is as compared to its ancestor lines
(previous versions of the line) and give full or partial credit to the
last author based on the analysis.