Add optimise timeline feature (#2180)
Add optimise timeline feature

Currently, the ramps have a lot of empty space if the since and until
dates are specified far apart. 

Let's add a checkbox that can eliminate this empty space for each ramp.
This is especially useful if RepoSense is to be used to set up a report
of an individual's past OSS contributions.