Align right panel with welcome title (#2229)
Align panel titles

The title text on the left panel is not flushed to the top because it
is a header.

There are several types of title texts on the right panel. Initially,
some were headers and others were not. There was padding to ensure that
those which were not headers were not flushed to the top.  However, this
meant that those headers were lower than expected and not aligned. 

To align the texts, let us remove the padding for the right panel and
set the first element on the right panel as a header.