[#2196] Use git blame line info for aggregate blame author modified and date info (#2232)
`aggregateBlameAuthorModifiedAndDateInfo` is hard to maintain, due to
complex string wrangling and seemingly magic numbers. Let's fix these using
`GitBlameLineInfo` introduced in #2140.

- Use blameLine to abstract away the string wrangling and improve
understandability. Magic numbers are also replaced to improve code quality.    

- In order to use blameLine however, it is also changed to use 
author-time instead of commit-time, as there is a discrepancy between the two,
causing some test cases to fail. 

 - The naming of the timestamp field in GitBlameLineInfo is changed to
reflect that it is in seconds, as author-time is in seconds.