[#1941] Remove test mode flag (#2314)
Remove test mode flag

Currently, the test mode flag is not utilised in our codebase. The test
mode flag was added to pass system tests due to legacy reasons and the
related logic was removed in a previous PR.

Let's remove the related code and documentation for the test mode flag
to avoid additional maintenance.