[#2332] Frontend Sort Function Bugfix and Enhancement (#2337)
Frontend Sort Function Enhancement

Improve frontend sort function efficiency for descending order by
refactoring to use boolean, and modify the sort comparator to take in 1
additional argument `isDesc` instead of calling `.reverse()` after
doing an ascending sort.


Important: A Cypress test case was modified to account for corrected
behavior because previous code for 'groupByAuthors' contains a bug
where we use `sortingWithinOption` instead of `sortWithinOption`, thus
causing the sort order within a group to be arbitrary if we sort by
'title' 
since the property doesn't exist and we are sorting an array of blank
strings.