[#2366] Support datetime since until (#2388)
Support datetime (previously just date) for since and until in both
Backend CLI and Frontend.

3 supported formats for since and until dates: 
1. DD/MM/YYYY
2. DD/MM/YYYY'T'HH:mm
3. DD/MM/YYYY'T'HH:mm:ss

For example:
`./gradlew run -Dargs="--since 31/12/2017T10:10"`


Rationale for using 'T' instead of blank space is to adhere to ISO 8601
format and prevent complicating the CLI input where it will need an
additional set of quotation marks to identify the space-separated
date.

Behavior: For since date, if hour/minute/seconds not supplied then
default to 0 for each. For until date, if hour/minute/seconds not
supplied then default to 23, 59, 59 respectively. If only one date
(since or until) is provided, and a period is also provided (default to
1 month in our existing project), the time of since is truncated to
00:00:00 if until is supplied, and time of until is truncated to
23:59:59 if since is supplied.

We use text input instead of datetime input only if the
detected Safari version is <14.1 to improve user experience for MacOS
users. 



Note for systemtest: For "Asia/Singapore" timezone, on dates before
01-01-1982, where Singapore and Malaysia switched from UTC+07:30 to
UTC+08:00, all date conversions will result in +7.5 hours instead of +8
hours. For example, the Unix epoch 1970-01-01 00:00:00 will be
1970-01-01 07:30:00 with "Asia/Singapore" timezone, but 1970-01-01
08:00:00 with "UTC+08" timezone.
Reference: https://en.wikipedia.org/wiki/UTC%2B07:30.