[#2210] Add default sort order for groups (#2331)
Add default sort order for groups in RepoSense. 

Now, we have 2 more options: default and default desc. It will order
the groups based on the initial order of the rows in repo-config.csv


Note: If we group by repos or group by authors, then the order among
groups will take the default rank of the first element in the group.
For example, if the groups to be sorted are: {[1, 6, 10], [4, 7, 9],
[3, 5], [2, 8]}, where the numbers are the default rank of the user
repos, then the final sorted groups will be {[1, 6, 10], [2, 8], [3,
5], [4, 7, 9]}